chatbox-api/requirements.txt
uvicorn[standard]
sqlalchemy>=1.4,<3
psycopg2-binary
pydantic>=1,<3
python-multipart==0.0.20
langchain==0.2.11
langchain-openai>="0.1.20"
python-dotenv>="1.0.1"
faiss-cpu==1.9.0.post1
fastapi==0.115.6
tiktoken>=0.7,<1
langchain-community==0.2.10
pypdf
python-docx
langdetect
langchain-postgres
psycopg
psycopg-binary
langchain-google-genai
langchain-openai